perf(bienbot-ui): extract memoised MessageContent component
Moves ENTITY_JSON_RE, PLACEHOLDER_RE, buildEntityChip, substituteChips,
processChildren, and renderMessageContent out of BienBotPanel.jsx into a
new React.memo component. The useMemo hook keys on text so the regex +
JSON.parse work runs once per message, not on every parent render.
